gnosh
git-svn-id: svn://132.239.158.177/projects/ratrix/trunk@1756 d291aab3-510b-4049-a6d1-fe65b8d52d70

+++ classes/@subject/subject.m
@@ -11,7 +11,7 @@ s.receivedDate=[];
 s.litterID='';
 s.supplier='';
 s.protocol=[];
-s.trainingStepNum=0;
+s.trainingStepNum=uint8(0);
 s.protocolVersion.manualVersion=0;
 
 switch nargin
