+++ setup/setupPhysRig1/setProtocolPhys.m
@@ -205,7 +205,7 @@ ts9 = trainingStep(vh, crf_trf_better,  repeatIndefinitely(), noTimeOff(), svnRe
 ts10= trainingStep(vh, rptUnq,  repeatIndefinitely(), noTimeOff(), svnRev); %rpt/unq
 
 p=protocol('practice phys',{ts1, ts2, ts3, ts4, ts5, ts6, ts7, ts8, ts9, ts10});
-stepNum=10;
+stepNum=uint8(10);
 
 for i=1:length(subjIDs),
     subj=getSubjectFromID(r,subjIDs{i});
